+++ src/Kubescape/globals.ts
@@ -1,3 +1,3 @@
-export const PACKAGE_STABLE_BUILD = "v2.0.144"
-export const ERROR_KUBESCAPE_NOT_INSTALLED = "Kubescape is not installed!"
-export const COMMAND_VIEW_CTRL_DOC = "kubescape.viewCtrlDoc"
+export const PACKAGE_STABLE_BUILD = "v2.0.176";
+export const ERROR_KUBESCAPE_NOT_INSTALLED = "Kubescape is not installed!";
+export const COMMAND_VIEW_CTRL_DOC = "kubescape.viewCtrlDoc";