+++ src/Kubescape/types.ts
@@ -1,10 +1,10 @@
 export class KubescapeVersion {
-    version: string
-    isLatest: boolean
+    version: string;
+    isLatest: boolean;
 
     constructor() {
-        this.version = "unknown"
-        this.isLatest = true
+        this.version = "unknown";
+        this.isLatest = true;
     }
 }
 
@@ -12,15 +12,15 @@ export type KubescapePath = {
     fullPath : string,
     baseDir : string,
     isCustom : boolean
-}
+};
 
 export type KubescapeFramework = {
     name : string,
     isInstalled : boolean,
     location : string
-}
+};
 
 export type KubescapeDir = {
     directory : string,
     isCustom : boolean
-}
+};