+++ src/extension.ts
@@ -1,30 +1,30 @@
 import * as vscode from 'vscode';
 
-import { IKubescapeConfig, KubescapeApi } from '@kubescape/install'
+import { IKubescapeConfig, KubescapeApi } from '@kubescape/install';
 
-import * as kubescape from './Kubescape/kubescape'
-import * as scan from './Kubescape/scan'
-import * as contextHelper from './utils/context'
-import { KubescapeConfig } from './Kubescape/config'
+import * as kubescape from './Kubescape/kubescape';
+import * as scan from './Kubescape/scan';
+import * as contextHelper from './utils/context';
+import { KubescapeConfig } from './Kubescape/config';
 import { Logger } from './utils/log';
-import { VscodeUi } from './utils/ui'
+import { VscodeUi } from './utils/ui';
 
 import {
 	ERROR_KUBESCAPE_NOT_INSTALLED
-} from './Kubescape/globals'
+} from './Kubescape/globals';
 import { KubescapeCodeAction } from './Kubescape/diagnostic';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export async function activate(context: vscode.ExtensionContext) {
-	contextHelper.setExtensionContext(context)
+	contextHelper.setExtensionContext(context);
 
-	let subscriptions = context.subscriptions
+	let subscriptions = context.subscriptions;
 
 	// Subscribe all the exported functions in kubescape.ts
 	for (let exportedFunc of Object.keys(kubescape)){
-		let fullFuncName = `kubescape.${exportedFunc}`
-		subscriptions.push(vscode.commands.registerCommand(fullFuncName, (args) => eval(`${fullFuncName}(args)`)))
+		let fullFuncName = `kubescape.${exportedFunc}`;
+		subscriptions.push(vscode.commands.registerCommand(fullFuncName, (args) => eval(`${fullFuncName}(args)`)));
 	}
 
 	subscriptions.push(
@@ -33,39 +33,39 @@ export async function activate(context: vscode.ExtensionContext) {
 		})
 	);
 
-	const kubescapeApi = KubescapeApi.instance
-	await initializeExtension(kubescapeApi)
+	const kubescapeApi = KubescapeApi.instance;
+	await initializeExtension(kubescapeApi);
 
 
 	if (!kubescapeApi.isInstalled) {
-		Logger.error(ERROR_KUBESCAPE_NOT_INSTALLED)
-		throw new Error
+		Logger.error(ERROR_KUBESCAPE_NOT_INSTALLED);
+		throw new Error;
 	}
 
 	/* Auto scan on save */
-	addOnSaveTextDocument(context)
+	addOnSaveTextDocument(context);
 
 	/* Scan on new file open */
-	addOnOpenTextDocument(context)
+	addOnOpenTextDocument(context);
 
 	/* Remove diagnostics on file close */
 	vscode.workspace.onDidCloseTextDocument(doc => {
-		scan.removeFileDiagnostics(doc.fileName)
-	})
+		scan.removeFileDiagnostics(doc.fileName);
+	});
 
 	/* First scan of current file */
 	if (vscode.window.activeTextEditor) {
-		const doc = vscode.window.activeTextEditor.document
-		scan.kubescapeScanYaml(doc)
+		const doc = vscode.window.activeTextEditor.document;
+		scan.kubescapeScanYaml(doc);
 	}
 
-	Logger.info("Kubescape in active")
+	Logger.info("Kubescape is active");
 }
 
 // this method is called when your extension is deactivated
 export function deactivate() {
-	contextHelper.setExtensionContext(undefined)
-	Logger.info("Kubescape deactivated")
+	contextHelper.setExtensionContext(undefined);
+	Logger.info("Kubescape deactivated");
 }
 
 function addOnSaveTextDocument(context : vscode.ExtensionContext) {
@@ -76,7 +76,7 @@ function addOnSaveTextDocument(context : vscode.ExtensionContext) {
 
 		if (KubescapeConfig.instance.scanOnSave) {
 			if (vscode.window.visibleTextEditors.some((e) => e.document.fileName === document.fileName)) {
-				scan.kubescapeScanYaml(document)
+				scan.kubescapeScanYaml(document);
 			}
 		}
 	}, null, context.subscriptions);
@@ -88,28 +88,28 @@ function addOnOpenTextDocument(context : vscode.ExtensionContext) {
 			return;
 		}
 
-		scan.kubescapeScanYaml(document)
+		scan.kubescapeScanYaml(document);
 	}, null, context.subscriptions);
 }
 
 async function initializeExtension(kubescapeApi : KubescapeApi) {
-    const config = KubescapeConfig.instance
+    const config = KubescapeConfig.instance;
 
 	await kubescapeApi.setup(new VscodeUi, new class implements IKubescapeConfig {
         get version() : string {
-            return config.kubescapeVersion
+            return config.kubescapeVersion;
         }
         get frameworksDirectory() : string | undefined {
-            return config.customFrameworkDirectory
+            return config.customFrameworkDirectory;
         }
         get baseDirectory(): string {
-            return config.kubescapeDir
+            return config.kubescapeDir;
         }
         get requiredFrameworks(): string[] | undefined {
-            return config.requiredFrameworks
+            return config.requiredFrameworks;
         }
         get scanFrameworks(): string[] | undefined {
-            return config.scanFrameworks
+            return config.scanFrameworks;
         }
-    })
+    });
 }