+++ src/utils/context.ts
@@ -1,7 +1,7 @@
-import * as vscode from 'vscode'
+import * as vscode from 'vscode';
 
 
-let extensionContext : vscode.ExtensionContext | undefined = undefined
+let extensionContext : vscode.ExtensionContext | undefined = undefined;
 
 export function setExtensionContext(context : vscode.ExtensionContext | undefined) {
     extensionContext = context;