+++ src/utils/extensionMethods.ts
@@ -1,4 +1,4 @@
-export {}
+export {};
 declare global {
     export interface String { 
         extractBetween(surround : string) : string
@@ -11,35 +11,35 @@ String.prototype.extractBetween = function(this: string, surround : string) {
     return this.substring(
         this.indexOf(surround) + 1,
         this.lastIndexOf(surround)
-    )
-}
+    );
+};
 
 String.prototype.toJson = function() : any {
-    let obj
-    let str = this.toString()
+    let obj;
+    let str = this.toString();
 
     try {
-        obj = JSON.parse(str)
+        obj = JSON.parse(str);
     } catch {
-        obj = {}
+        obj = {};
     }
 
-    return obj 
-}
+    return obj; 
+};
 
 String.prototype.toJsonArray = function() : any {
-    let obj
-    let str = this.toString()
+    let obj;
+    let str = this.toString();
 
     try {
         if (this[0] !== '[') {
-            str = '[' + str + ']'
+            str = '[' + str + ']';
         }
 
-        obj = JSON.parse(str)
+        obj = JSON.parse(str);
     } catch {
-        obj = []
+        obj = [];
     }
 
-    return obj 
-}
+    return obj; 
+};